+++ shared/src/index.ts
@@ -14,6 +14,7 @@ export * from "./model/tableStyles";
 export * from "./model/ops";
 export * from "./model/sections";
 export * from "./model/query";
+export * from "./model/documentEditor";
 export * from "./model/sdt";
 export * from "./model/math";
 export * from "./model/tableGrid";